refactor(tests): reorganize test directory structure
Reorganize test directory structure by:
- Moving camera tests from test_scenes/camera to camera/
- Moving primitive and light config files to their own directories
- Adding separate test files for primitives and lights
- Removing test.cpp and splitting functionality into dedicated test
files

This change improves organization and maintainability of the testing
directory structure.
