+++ CMakeLists.txt
@@ -318,6 +318,8 @@ if(BUILD_TESTING)
          tests/test_scenes/camera/parseCamera.cpp
          tests/test_scenes/camera/missingFields/missingFields.cpp
          tests/test_scenes/camera/missingFields/resolution/missingResolutionFields.cpp
+         tests/test_scenes/primitives/parsePrimitives.cpp
+         tests/test_scenes/lights/parseLights.cpp
          src/ParserConfigFile.cpp
          src/Factory.cpp
          src/lights/LightComposite.cpp
