+++ tests/camera/missingFields/missingFOV.cfg
