+++ tests/camera/missingFields/missingPosition.cfg
