+++ tests/camera/missingFields/missingResolution.cfg
