+++ tests/camera/missingFields/missingRotation.cfg
