+++ tests/camera/missingFields/position/missingX.cfg
