+++ tests/camera/missingFields/position/missingY.cfg
