+++ tests/camera/missingFields/position/missingZ.cfg
