+++ tests/camera/missingFields/resolution/missingHeight.cfg
