+++ tests/camera/missingFields/resolution/missingWidth.cfg
