+++ tests/camera/missingFields/rotation/missingX.cfg
