+++ tests/camera/missingFields/rotation/missingY.cfg
