+++ tests/camera/missingFields/rotation/missingZ.cfg
