+++ tests/camera/parseCamera.cfg
