+++ tests/invalidConfig.cfg
@@ -21,7 +21,7 @@ primitives :
 {
   spheres = (
     {
-      x = 0.0;;
+      x = 0.0;; # Extra semicolon
       y = 0.0;
       z = -1.0;
       r = 0.5;
