+++ tests/lights/parseLights.cfg
