+++ tests/lights/parseLights.cpp
@@ -22,20 +22,21 @@ class ParserConfigFileTest : public ::testing::Test {
     std::vector<std::string> _plugins;
 };
 
-TEST_F(ParserConfigFileTest, MissingCameraResolutionFields) {
-  _cfgFile = "tests/test_scenes/camera/missingFields/resolution/missingWidth.cfg";
+TEST_F(ParserConfigFileTest, TestParseLights) {
+  _cfgFile = "tests/lights/parseLights.cfg";
   Raytracer::ParserConfigFile parser(_cfgFile, _plugins);
   Raytracer::Camera camera;
   Raytracer::ShapeComposite sc;
   Raytracer::LightComposite lc;
 
-  EXPECT_EQ(camera.getWidth(), 0);
+  try {
+    parser.parseConfigFile(camera, sc, lc);
+  } catch (const Raytracer::ParseError &e) {
+    FAIL() << "Failed to parse lights configuration: " << e.what();
+  } catch (const Raytracer::ConfigError &e) {
+    FAIL() << "Failed to parse lights configuration (ConfigError): "
+           << e.what();
+  }
 
-  _cfgFile = "tests/test_scenes/camera/missingFields/resolution/missingHeight.cfg";
-  Raytracer::ParserConfigFile parser2(_cfgFile, _plugins);
-  Raytracer::Camera camera2;
-  Raytracer::ShapeComposite sc2;
-  Raytracer::LightComposite lc2;
-
-  EXPECT_EQ(camera2.getHeight(), 0);
+  EXPECT_EQ(lc.getLights().size(), 2);
 }
