+++ tests/primitives/parsePrimitives.cfg
